fix: fix some bugs

+++ MagicEventsWebApp/Backend/events-management-service/src/main/java/com/service/eventsmanagementservice/controller/EventGestorController.java
@@ -327,7 +327,13 @@ public class EventGestorController {
     }
 
     @DeleteMapping("/deletepartecipant")
-    public boolean deleteUser(@RequestParam("magicEventsTag") Long magicEventsTag){
+    public boolean deleteUser(
+            @RequestHeader("Authorization") String authorizationHeader,
+            @RequestParam("magicEventsTag") Long magicEventsTag
+    ){
+        if (!tokenValidatorService.isTokenValid(extractToken(authorizationHeader))) {
+            return false;
+        }
         return eventGestorService.deleteUser(magicEventsTag);
     }
